+++ tests/TestCase.php
@@ -84,6 +84,10 @@ abstract class TestCase extends BaseTestCase
             {
             }
 
+            public function removeDirectory(Server $server, string $remotePath, bool $sudo = false): void
+            {
+            }
+
             public function runCommand(Server $server, string $command, bool $sudo = false): void
             {
             }